feat: enhance AllergyController and UserProfileController with command services for improved allergy and user profile management

+++ services/profiles-service/src/main/java/pe/edu/upc/profiles_service/profiles/interfaces/rest/transform/UserProfileResourceFromEntityAssembler.java
@@ -10,7 +10,7 @@ public class UserProfileResourceFromEntityAssembler {
 
     public static UserProfileResource toResourceFromEntity(UserProfile entity) {
         return new UserProfileResource(
-                entity.getId(),
+                entity.getId().intValue(),
                 entity.getGender(),
                 entity.getHeight(),
                 entity.getWeight(),
